Now when Advertisements are created, author ID is taken from the token.

+++ src/main/java/com/example/main/dto/AdvertisementDto.java
@@ -14,7 +14,7 @@ public abstract class AdvertisementDto {
     @NonNull
     private String description;
     @NonNull
-    private int authorId;
+    private Long authorId;
 
     private String authorUsername;
     private String authorEmail;
