+++ src/main/java/com/example/main/dto/OfferDto.java
@@ -1,7 +1,7 @@
 package com.example.main.dto;
 
 public class OfferDto extends AdvertisementDto {
-    public OfferDto(Long id, String title, String description, int authorId) {
+    public OfferDto(Long id, String title, String description, Long authorId) {
         super(id, title, description, authorId);
     }
 }
