+++ src/main/java/com/example/main/dto/RequestDto.java
@@ -1,7 +1,7 @@
 package com.example.main.dto;
 
 public class RequestDto extends AdvertisementDto {
-    public RequestDto(Long id, String title, String description, int authorId) {
+    public RequestDto(Long id, String title, String description, Long authorId) {
         super(id, title, description, authorId);
     }
 }
