+++ src/main/java/com/example/main/model/Advertisement.java
@@ -27,7 +27,7 @@ public class Advertisement {
     @NonNull
     private String description;
     @NonNull
-    private int authorId;
+    private Long authorId;
     @NonNull
     private double lat;
     @NonNull
